refactor(crud-votacao): adicionado o enum CpfStaus dentro do pacote enum para padronizacao do codigo
