+++ src/main/java/com/faculdade/votacao/controller/CpfValidatorController.java
@@ -4,8 +4,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import com.faculdade.votacao.enums.CpfStatus;
 import com.faculdade.votacao.service.CpfValidatorService;
-import com.faculdade.votacao.service.CpfValidatorService.CpfStatus;
 
 import java.util.HashMap;
 import java.util.Map;
