+++ src/main/java/com/faculdade/votacao/enums/CpfStatus.java
@@ -1,0 +1,6 @@
+package com.faculdade.votacao.enums;
+
+public enum CpfStatus {
+    ABLE_TO_VOTE,
+    UNABLE_TO_VOTE
+}
