+++ src/main/java/com/faculdade/votacao/interfaces/CpfValidatorInterface.java
@@ -1,6 +1,6 @@
 package com.faculdade.votacao.interfaces;
 
-import com.faculdade.votacao.service.CpfValidatorService.CpfStatus;
+import com.faculdade.votacao.enums.CpfStatus;
 
 public interface CpfValidatorInterface {
     CpfStatus validarCpf(String cpf);
